pyproject.toml
[project]
name = "kdctl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "kubernetes>=33.1.0",
    "pyyaml>=6.0.2",
    "typer>=0.16.0",
]

[project.scripts]
kdctl = "kdctl.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
